refactor(connect): move monero method definitions to corresponding folder
