+++ packages/connect-common/src/types/api/monero/moneroGetAddress.ts
@@ -1,0 +1,7 @@
+import type { MoneroAddress, MoneroGetAddress } from './common';
+import type { BundledParams, Params, Response } from '../../params';
+
+export declare function moneroGetAddress(params: Params<MoneroGetAddress>): Response<MoneroAddress>;
+export declare function moneroGetAddress(
+    params: BundledParams<MoneroGetAddress>,
+): Response<MoneroAddress[]>;
