+++ packages/connect-common/src/types/api/monero/moneroGetWatchKey.ts
@@ -1,0 +1,6 @@
+import type { MoneroGetWatchKey, MoneroWatchKey } from './common';
+import type { Params, Response } from '../../params';
+
+export declare function moneroGetWatchKey(
+    params: Params<MoneroGetWatchKey>,
+): Response<MoneroWatchKey>;
