+++ packages/connect-common/src/types/api/monero/moneroKeyImageSync.ts
@@ -1,0 +1,6 @@
+import type { MoneroKeyImageSync, MoneroKeyImageSyncResult } from './common';
+import type { Params, Response } from '../../params';
+
+export declare function moneroKeyImageSync(
+    params: Params<MoneroKeyImageSync>,
+): Response<MoneroKeyImageSyncResult>;
