+++ packages/connect-common/src/types/api/monero/moneroSignTransaction.ts
@@ -1,0 +1,6 @@
+import type { MoneroSignTransaction, MoneroSignedTransaction } from './common';
+import type { Params, Response } from '../../params';
+
+export declare function moneroSignTransaction(
+    params: Params<MoneroSignTransaction>,
+): Response<MoneroSignedTransaction>;
