--- packages/connect-common/src/types/api/moneroGetWatchKey.ts
@@ -1,6 +1,0 @@
-import type { Params, Response } from '../params';
-import type { MoneroGetWatchKey, MoneroWatchKey } from './monero';
-
-export declare function moneroGetWatchKey(
-    params: Params<MoneroGetWatchKey>,
-): Response<MoneroWatchKey>;
