--- packages/connect-common/src/types/api/moneroKeyImageSync.ts
@@ -1,6 +1,0 @@
-import type { Params, Response } from '../params';
-import type { MoneroKeyImageSync, MoneroKeyImageSyncResult } from './monero';
-
-export declare function moneroKeyImageSync(
-    params: Params<MoneroKeyImageSync>,
-): Response<MoneroKeyImageSyncResult>;
