--- packages/connect-common/src/types/api/moneroSignTransaction.ts
@@ -1,6 +1,0 @@
-import type { Params, Response } from '../params';
-import type { MoneroSignTransaction, MoneroSignedTransaction } from './monero';
-
-export declare function moneroSignTransaction(
-    params: Params<MoneroSignTransaction>,
-): Response<MoneroSignedTransaction>;
